refactor: create components directory
