--- src/TextMotion/TextMotion.scss
@@ -1,41 +1,0 @@
-@keyframes fadeIn {
-  from {
-    opacity: 0;
-  }
-  to {
-    opacity: 1;
-  }
-}
-
-@keyframes fadeOut {
-  from {
-    opacity: 1;
-  }
-  to {
-    opacity: 0;
-  }
-}
-
-@keyframes slideUp {
-  from {
-    transform: translateY(0.4rem);
-  }
-  to {
-    transform: translateY(0);
-  }
-}
-
-@keyframes slideDown {
-  from {
-    transform: translateY(-0.4rem);
-  }
-  to {
-    transform: translateY(0);
-  }
-}
-
-.text-motion span {
-  display: inline-block;
-  white-space: pre;
-  animation: var(--animation) var(--duration) ease-out var(--delay) both;
-}
