+++ src/components/TextMotion/TextMotion.tsx
@@ -1,15 +1,18 @@
 import './TextMotion.scss';
+import '../../styles/animations.scss';
 
 import React, { useMemo } from 'react';
 
+import { splitText } from '../../utils';
+
+export type SplitType = 'character' | 'word';
 type PresetType = 'fadeIn' | 'fadeOut' | 'slideUp' | 'slideDown';
-type SplitType = 'character' | 'word';
 
 type TextMotionProps = {
   as?: keyof React.JSX.IntrinsicElements;
   text: string;
-  presets?: PresetType[];
   split?: SplitType;
+  presets?: PresetType[];
 };
 
 const DEFAULT_DURATION = 0.25;
@@ -22,13 +25,7 @@ const generateAnimationByPresets = (presets: PresetType[], index: number): React
   return { animation };
 };
 
-export const splitText = (text: string, split: SplitType): string[] => {
-  if (split === 'character') return text.split('');
-  if (split === 'word') return text.split(/(\s+)/);
-  return text.split('');
-};
-
-export const TextMotion = ({ as = 'span', text, presets = ['fadeIn'], split = 'character' }: TextMotionProps) => {
+export const TextMotion = ({ as = 'span', text, split = 'character', presets = ['fadeIn'] }: TextMotionProps) => {
   const Tag = as;
   const letters = useMemo(() => splitText(text, split), [text, split]);
